refactor Trainer
