+++ pytorch_lightning/accelerators/dp_accelerator.py
@@ -16,7 +16,6 @@ from typing import Optional, Union
 import torch
 from torch import optim
 
-from pytorch_lightning import _logger as log
 from pytorch_lightning.accelerators.accelerator import Accelerator
 from pytorch_lightning.cluster_environments import ClusterEnvironment
 from pytorch_lightning.core.lightning import LightningModule
